added support for multiple keys
Now you can make more than one key (by trying to encode/decode a
document with a non-existent key), select which key to use (with the
optional key_name argument at the end of the Terminal command), and
view a list of all your key names. This allows you to add security by
encoding documents with multiple keys (just make sure you remember to
decode the document in the reverse order of how you encoded it!).